refactor: remove internal prefix from src files
